stb0899 - Spectrum Scan improved.
By setting the AGCRFCFG register to what is used in stb0899_search() for DVB-S,
in get_spectrum_scan(). No longer necessary to attempt to tune a DVB-S signal
to get a 1/2 way decent looking spectrum plot.
Not bad for 20+ year old technology at least.

Use IQSYMBSEL in get_constellation_samples(), to try and allow monitoring from
different stages. Still unable to plot from the DVB-S2 demod.

Modify ES_QUANT again, as cnr reports are still out of whack on dvb-s2.
May need some more testing on diff TPs and in depth analysis to get it right.
But this seems better than original which was always reporting cnr in the 17 dB range.
Found this note in datasheet...
ESNO_EST : signal-to-noise ratio estimate (to convert to dB use:
 -10 log (ESNO_EST / ESNO_QUANT ^ 2)

Signed-off-by: Bill Murphy <gc2majortom@gmail.com>