Moved drivers/base/firmware_loader/builtin firmware to firmware/ Updated README
A few years ago upstream stopped providing linux-firmware that was
located in firmware/ with the kernel. Their reason for doing this
is all major distro's provide and update the linux-firmware package
regularly.

At that time UDL then started merging in the linux-firmware tree
but changed the target directory to drivers/base/firmware_loader/builtin
to put the firmware. There has been an on going git merge issue that
will not go away involving a Makefile in this directory. I do not
believe the existing Makefile is being used at this time but
there is always the possibility it could be in the future.

Although most of the firmware is redundant because of it getting
installed by the linux-firmware package there are some old firmware
present that is not in the linux-firmware package. Also there are
some dvb/atsc/dvbt related firmware that is not present in the
linux-firmware package.

The best thing would be to remove what is in the linux-firmware package
to stream line UDl's repo but I do not have the time to do it.